Move legacy runtime adapter into app core
